app/requirements.txt
tensorflow
pandas
numpy
opencv-python
fastapi
uvicorn
python-multipart
streamlit
requests
Pillow
fpdf